refactor: remove SlashIDLoaded barrel file
